JDK-8254627: Cleanup {Abstract,Single,Split}IndexWriter classes
