+++ src/jdk.javadoc/share/classes/jdk/javadoc/internal/doclets/formats/html/resources/standard.properties
@@ -74,9 +74,12 @@ doclet.navDeprecated=Deprecated
 doclet.Window_Deprecated_List=Deprecated List
 doclet.Overrides=Overrides:
 doclet.in_class=in class
+doclet.Element_in=Element in {0}
+doclet.Enum_constant_in=Enum constant in {0}
 doclet.Static_variable_in=Static variable in {0}
 doclet.Variable_in=Variable in {0}
 doclet.Constructor_for=Constructor for {0}
+doclet.Record_component_in=Record component of {0}
 doclet.Static_method_in=Static method in {0}
 doclet.Search_tag_in=Search tag in {0}
 doclet.Method_in=Method in {0}
