+++ src/jdk.javadoc/share/classes/jdk/javadoc/internal/doclets/toolkit/resources/doclets.properties
@@ -163,6 +163,7 @@ doclet.interfaces=interfaces
 doclet.class=class
 doclet.classes=classes
 doclet.Record=Record
+doclet.record=record
 doclet.Error=Error
 doclet.error=error
 doclet.errors=errors
